Add files via upload
This is one of the complete projects I've made. it is a banking system but (for all bank owners and investors. the system is not adaptable yet :), but I bet that it would be an upgrade for Al Rajhi Bank! (( haha just joking with them since I'm their loyal client so I think I earned the rights to joke about my bank)